pyproject.toml
[project]
name = "particulars-setup"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "annotated-types==0.7.0",
    "bidict==0.23.1",
    "certifi==2024.7.4",
    "charset-normalizer==3.3.2",
    "contourpy==1.2.1",
    "cycler==0.12.1",
    "fonttools==4.53.1",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "grpcio==1.74.0",
    "idna==3.7",
    "influxdb==5.3.2",
    "intelhex==2.3.0",
    "kiwisolver==1.4.5",
    "kneed==0.8.5",
    "llvmlite==0.43.0",
    "matplotlib==3.9.1",
    "mpmath==1.3.0",
    "msgpack==1.0.8",
    "msl-loadlib==0.10.0",
    "numba==0.60.0",
    "numpy==1.26.4",
    "packaging==24.1",
    "pandas==2.2.2",
    "pillow==10.4.0",
    "protobuf==5.29.5",
    "psutil==6.0.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyparsing==3.1.2",
    "pyserial==3.5",
    "pyside6==6.7.2",
    "pyside6-addons==6.7.2",
    "pyside6-essentials==6.7.2",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.1",
    "pyvisa==1.14.1",
    "requests==2.32.3",
    "scipy==1.14.0",
    "seaborn==0.13.2",
    "sensirion-driver-adapters==2.1.9",
    "sensirion-driver-support-types==0.2.0",
    "sensirion-i2c-driver==1.0.0",
    "sensirion-shdlc-driver==0.1.5",
    "sensirion-shdlc-sensorbridge==0.2.0",
    "shiboken6==6.7.2",
    "six==1.16.0",
    "smmap==5.0.2",
    "sympy==1.13.1",
    "tekhsi==1.0.0",
    "tm-data-types==0.1.1",
    "typing-extensions==4.12.2",
    "typing-inspection==0.4.1",
    "tzdata==2024.1",
    "urllib3==2.2.2",
]
